pagination duplicating via template + other changes
White space wrap for menu items on top,
margin right for navigation,
less transparency for "top" but still not a very very contrasting.. why? bc it can not compete with the rest of the site. This is not a "primary" navigation item, only a helper and there is already tooo much happening here